+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/exporters/modes/FrameExportMode.java
@@ -57,12 +57,16 @@ public enum FrameExportMode {
      */
     WEBP,
     /**
+     * WEBP - animated
+     */
+    WEBP_ANIMATED,
+    /**
      * SWF - Shockwave Flash
      */
     SWF;
     
     public boolean available() {
-        if (this == WEBP) {
+        if (this == WEBP || this == WEBP_ANIMATED) {
             return ImageFormat.WEBP.available();
         }
         return true;