+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/exporters/modes/MorphShapeExportMode.java
@@ -24,41 +24,100 @@ import com.jpexs.decompiler.flash.tags.enums.ImageFormat;
  * @author JPEXS
  */
 public enum MorphShapeExportMode {
-    //TODO: implement other morphshape export modes
-
     /**
      * SVG animation - Scalable Vector Graphics
      */
-    SVG,
+    SVG(true, false),
     /**
      * SVG start and end frames - Scalable Vector Graphics
      */
-    SVG_START_END,
+    SVG_START_END(false, false),
+    /**
+     * SVG individual frames - Scalable Vector Graphics
+     */
+    SVG_FRAMES(false, true),
     /**
      * HTML5 canvas animation
      */
-    CANVAS,
+    CANVAS(true, false),
     /**
      * PNG start and end frames - Portable Network Graphics
      */
-    PNG_START_END,
+    PNG_START_END(false, false),
+    /**
+     * PNG individual frames - Portable Network Graphics
+     */    
+    PNG_FRAMES(false, true),
     /**
      * BMP start and end frames - Windows Bitmap
      */
-    BMP_START_END,
+    BMP_START_END(false, false),
+    /**
+     * BMP individual frames - Windows Bitmap
+     */    
+    BMP_FRAMES(false, true),
     /**
      * WEBP start and end frames
      */
-    WEBP_START_END,
-    //GIF,
-    //AVI,
+    WEBP_START_END(false, false),
+    /**
+     * WEBP individual frames
+     */
+    WEBP_FRAMES(false, true),
+    /**
+     * GIF - Graphics Interchange Format
+     */
+    GIF(true, false),
+    /**
+     * AVI - Audio Video Interleave
+     */
+    AVI(true, false),
     /**
      * SWF - Shockwave Flash
      */
-    SWF;
+    SWF(true, false),
+
+    /**
+     * WEBP
+     */
+    WEBP(true, false);
     
+    /**
+     * Whether this mode requires total duration in seconds
+     */
+    private final boolean duration;
+    /**
+     * Whether this mode requires number of frames
+     */
+    private final boolean frames;
+
+    private MorphShapeExportMode(boolean duration, boolean frames) {
+        this.duration = duration;
+        this.frames = frames;
+    }
+            
+    /**
+     * Checks whether this mode requires total time (length) in seconds.
+     * @return True on required
+     */
+    public boolean hasDuration() {
+        return duration;
+    }
+    
+    /**
+     * Checks whether this mode requires number of frames.
+     * @return True on required
+     */
+    public boolean hasFrames() {
+        return frames;
+    }
+    
+    /**
+     * Checks whether this mode is available on current platform.
+     * @return True on available.
+     */
     public boolean available() {
-        if (this == WEBP_START_END) {
+        if (this == WEBP_START_END || this == WEBP_FRAMES || this == WEBP) {
             return ImageFormat.WEBP.available();
         }
         return true;