+++ src/com/jpexs/decompiler/flash/gui/PreviewPanel.java
@@ -2144,7 +2144,7 @@ public class PreviewPanel extends JPersistentSplitPane implements TagEditorPanel
 
                 SWFHeader header;
                 try (OutputStream fos = new BufferedOutputStream(new FileOutputStream(extTempFile))) {
-                    header = new PreviewExporter().exportSwf(fos, treeItem, backgroundColor, fontPageNum, true);
+                    header = new PreviewExporter().exportSwf(fos, treeItem, backgroundColor, fontPageNum, true, PreviewExporter.MORPH_SHAPE_DEFAULT_DURATION);
                 }
             }
             Main.runAsync(extTempFile);
@@ -2182,7 +2182,7 @@ public class PreviewPanel extends JPersistentSplitPane implements TagEditorPanel
 
             SWFHeader header;
             try (OutputStream fos = new BufferedOutputStream(new FileOutputStream(tempFile))) {
-                header = new PreviewExporter().exportSwf(fos, treeItem, backgroundColor, fontPageNum, false);
+                header = new PreviewExporter().exportSwf(fos, treeItem, backgroundColor, fontPageNum, false, PreviewExporter.MORPH_SHAPE_DEFAULT_DURATION);
             }
 
             this.currentItem = treeItem;